kafka-spark-docker/producer/requirements.txt
kafka-python==2.0.2
watchdog==4.0.1